Fix notifications in CWS being translated after being formatted
We would mark for translation the message with placeholders (which is
correct) but we would look the message up in the catalog after having
formatted it, and then of course we wouldn't find any translation.

This also fixes an issue with message extraction.

Backport of fa15aa1af692fd1a4d8cfc04146b7a91e68583f2 from upstream/master.