Fix submissions_status RPC when UNION ALL doesn't keep the order (#1201)
UNION ALL is not guaranteed to keep the order of the results. In fact
with (at least) Postgres 13 the order is not well defined and can change
between executions.

This patch adds an extra column to the query, tagging each statistics
with its key, so that we can reconstruct the correct order of the
results.